pyproject.toml
[tool.poetry]
name = "pomo"
version = "0.0.1"
description = "Simple CLI Pomodoro timer"
authors = ["Your Name <you@example.com>"]

[tool.black]
line-length = 100
target_version = ["py37"]

[tool.poetry.dependencies]
python = "^3.7"
tqdm = "^4.45.0"
pydub = "^0.23.1"
fire = "^0.3.1"
python-dotenv = "^0.13.0"
loguru = "^0.4.1"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
pomo = "pomo.pomo:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
